pyproject.toml
[project]
name = "openenv-data-agent"
version = "0.1.0"
description = "Data Analysis OpenEnv Environment"
authors = [{ name = "Dhruba Saha" }]
dependencies = [
"pandas",
"numpy",
"openai",
"python-dotenv",
"openenv-core>=0.2.0",
"fastapi",
"uvicorn"
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
